modify demo and add test
