+++ examples/next/visual-tests/js/demo/src/datagrid.js
@@ -62,7 +62,7 @@ const root = document.getElementById('root');
 const example = document.createElement('div');
 root.appendChild(example);
 
-export function initializeDataGrid() {
+export function initializeTwoTablesDemo() {
   new Handsontable(example, {
     data: generateExampleData(),
     layoutDirection: isArabicDemoEnabled() ? "rtl" : "ltr",
