+++ visual-tests/playwright-e2e.config.ts
@@ -21,7 +21,7 @@ const config: PlaywrightTestConfig = {
     timeout: 10000
   },
   /* Run tests in files in parallel */
-  fullyParallel: false,
+  fullyParallel: true,
   /* Fail the build on CI if you accidentally left test.only in the source code. */
   forbidOnly: !!process.env.CI,
   /* Retry on CI only */
