+++ visual-tests/tests/e2e/copy-paste.spec.ts
@@ -4,7 +4,7 @@ import { helpers } from '../../src/helpers';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 testE2E(__filename, async({ tablePage }) => {
-  await tablePage.goto('/scenario-grid');
+  await tablePage.goto('/two-tables-demo');
 
   const tableTop = tablePage.locator('#tableTop > .handsontable');
   const tableBottom = tablePage.locator('#tableBottom > .handsontable');
